Fixed some of my If Statements to be shorter with the ! instead.
Still can't figure out why Insurance cost are not being charged on
Launch, but are being returned on Recycle.. Both these fuctions use the
same properties.. And they all use nathans new parts code.  So the
crew() is working because we are getting the insurance back.. But its
not working on launch.. Or the window were it shows the cost..
Completely lost on this one for now.. Suggest no insurnace cost for now
Set it to 0.