Added more KerbalNaut Stuff.
Starting to look around if there is a way to find out if the player has
hired a new kerbalnaut.. But I'm stating to think that the Access to
This new Part of Kerbal Space Program is still very limited.. Or just
plain undocumented.  At this point I feel that the first release will
have no interaction at all with the New Kerbal Training Building.. And
will be an option only for the player to train a kerbal within MC
intself.