Changed Some Of The DockingGoal To see If It works better
My test have shown writing the docking goal using the ingame Event
docking seems to work.. No idea if it works better or not.  Left most of
the old code IE the fallback code in place in case it fails.  So it
works just like the old version.. Only replaced the one part with ingame
version of docked.