Setting up the new RepairGoal Has Problems though
Worst problem is that it doesn't work at all.  I tried it so many ways
just a simple Bool being changed to True.. Nope.. Then I went the hard
root copied the DockingGoal almost word for word in every section.
Added a flag for Repair.. And even though everything works and the Bool
changes to True flag sets to REPAIR.. And Gets Reset back to NOTHING..
The god dam Mission Goal STILL WILL NOT COMPLETE.

The way the mission goal copletes is that the player in EVA right clicks
the part and selects Repair.  Once this is done. The Flag Is Set, and
The Bool Is changed to True.. (this all happens in game)  But for some
reason it will not read is complete the mission goals.. I give up for
now.. Wasted way to much time on this today.