The New Repair Part For MC (just place holder for now)
Right now in Mission Controller even though RepairGoal Does not work its
easy to set up and test.. Just place part on vessel.. Launch and land it
on ground.  Then right click the part (kerbal inside) and select repair.
Future version you will have to be EVA (that part does work but was
removed to help test the pain in the ass this has become.

The part is just a copied LargeBattery and is located in utilities