Ok I know it works.. But it seems its the context button that problem
I got the missiongoal to be accepted with the GUI Button thats on the
Main GUI. But the same method does not work for the Context Menu.  So
for some reason the context menu and Mission Controller goals don't like
each other.  I might have to do something else to get this to work
correctly.