more changes and still will not work.. This is frustrating to the end!!
GUI Button is disabled right now. But the Context menu is setup just
like all the other mods now.. Nothing more to add to the code to make
them work.. The context menu does work.. It does reach the 2nd part of
code to set flag to REPAIR.. (which is part of missioncontrollerEvents)
so it is changing that part.. But it stops cold there!!  If you use the
GUI Button version the whole code is executed and the Mission Is
Complete.  This is beyond my abilities to figure out at  this time..
putting this on hold.. Maybe Nathan can figure it out.  (has to be
permissions or something)