Yes Got It to Work.. Read for my mistake
Yes the goal now works.. My mistake.. And it was such a simple mistake
was that i did not use the Static Keyword to make the bool not have to
be initialized in a seperate Class... was such a dumb mistake... But I
am new to c# so I guess I can let it pass. ;)

Repair Goal Now works..  Still has some other stuff I need to work on to
make it better.. But it works if you make a manual mission.

Future Goals..

Make the goal random and be able to check for vessels with the part
already on them and make random missions to that vessel for repair. (the
hard part hahaha)