More changes To Undock and Dock
Had to mess around with Undock and Dock some more. Because of the way
KSP handles Vessel IDs when docked.. One of the biggest reason that Dock
had issues was because sometimes KSP would choose the Wrong Vessel ID to
assign to The Docking Vessel... (not all the time).

With Undock this problem became a million fold.. With the way KSP saves
GoalStatus when you undocked with the new Undock Comman the Vessel that
was ID in GoalStatus was always the wrong one.. IE the Space Station.
(you can tell this by how you have to always switch to the undocked
vessel after undocking).. Caused all kinds of issues.

So I came up with a new method.. And it also saves to the GoalStatus..
Instead of saving the vessel Id to the Goal for Dock And Undock I only
saved the Mission ID.  This way the goal still plants.. And you won't
have a forever True Statement with Dock and Undock.  The only isssue
still not tested is if Dock and Undock can be used with a
SubmissionGoal... Still have to test this out.