I returned docking back to record IDs
I thought I might of fixed a problem with the Docking And IDs.. And I
did.. But it introduced a new problem.. You could not have more then 1
docking port on a vessel.

What this means now.. For a normal mission docking works the way it
always has worked!

But if you want to use Undocking for a mission.. And you have to Dock at
the end of the mission then you need to use the value special = true for
the docking goal.  If you don't then Undocking will work and docking
won't.. At least with my test.