New RepairGoal Is Finished
RepairGoal has its own timer now.  Uses unity countdown.. But the time
is not perfect.. Its not really in seconds..But it almost looks like it.
Uses a float value of 1.0F. Then counts down.  Takes about a minute to
do the whole countdown.

The countdown is included in the display of the RepairGoal while in
mission for player can see it.