PreDevBuild
Have Most of the systems working for contracts.
Still missing Company info and Icons.
But until I can find a realiable random Selection for List to pick the
names from, this fuction will not work.

I used the standard random for the Final Ship List for Repair Parts..
But it only selects one ship name.. The top name.  So I might have to
write a new random selection program for this.  Simular to the Weighted
system for selection of Contract Missions.