DevBuild1
Had a nasty bug that took most of the day to find.
Set a value for KSPstartup to false.  So MCE did not liike the fact that
it did not get restarted after leaving flight.  and would start doing
crazy things like double charging for takeoffs. All better now and
things seem to be working. Almost ready for release.