Added Vessel Independent to Custom Contracts
Fixed the problem with using seperate large vessels in a mission.
Before if you put together a large vessel by piece.  Then used the
vessel in a mission, parts of the mission would not be able to be
completed because all the vessels have different IDs.  Only useful in
this situation.  If a vessel is launched all in one piece this is not
needed at all so keep it false!

Added more info to the Custom Contracts Display screen.  Made it more of
a manual.