Update .54 Testing Version See changelog
1. Added new Pop Up menu when Player Crashes vessel (gives option to
Delete Current Completed MissionGoals if doing a current mission)  Good
if you want to accept death and use a new vessel.
2. Added new Pop up for when you get a SubMission Payout.
3. Fixed a few Null refrence bugs in code that showed in Debug Menu..
4. No longer have the Current Budget Tab in the Space Center View (one
of the error codes, Space Center Don't like it)
5. Added some more values to the API.  Load Save, Load Backup, and Save.
Gives other mods the ability to Load an older save of MCE (if having
there own revert) Load a Current Save (for whatever reason) and ability
to save MCE if need be.
6.  Added an Option to settings for player to choose KSP Skin or Smoke
Skin.
7. Cleaned up some code that was not being used anymore in MCE.