Experimental Build 2
Changes so Far.
Fixed the Goal Payment popup coming popping up and staying.  Now if it
does show up for some strange reason it will go away once a mission is
deselected.

Changed the TIME and seconds countdown.  In older version these would
not be persitent if you left the vessel.. Ie they would start over. Now
if you leave vessel, or game they will be saved. So when you restart
again the Time Countdown will continue.  If you go and do something else
the time will continue to countdown for the Vessel the mission is
assigned too.  Please note though, that you can only do 1 Timed mission
at a time. If you Start another Timed mission the timer will be set for
the new mission and you will lose the old timer.

Added 1st of 3 New Mission Goals.  RoverResearch.  Comes with a new part
and will allow you to do science with rovers and or Probe Landers on any
bodies. Most of these types of missions are part of the Contracts Rover
Missions in the Random Contracts.  (WIP)