Worked on Realchutes values.  Read description
So the realChutes are kind of a challenge to balance out with the
regular Vanilla chutes. This is how it works.

Realchutes has value called deployed area.
Realchutes has value called dragCoefficient.

Total Drag of chutes is calculated by DragCoefficent * Deployed area.

Then Total Drag is calculated into Part Drag. so Part Drag * Total Drag
= Part Drag.

Then a comparison is made its the same as Vanilla chutes were looking
for a Total Parts Mass * Parachute drag Per ton (which is 70 default)
and if this value is Less than Part Drag.. Then that piece of vessel
will count toward Recycle.

For instance.

A RT-10 Solid Fuel booster with a Mk16 Parachute (realchutes) gives us
these values.
Area 490.9379
DragC 1
Total Drag = 490.9379
EveryPart on the Ship will be calculated toward mass.
So partmass * 490.9379. in this case it averaged out at 63.06

After all those calcuations we do the check.

So total mass of RTTank and Chute is .64 * 70 = 45.74
is 45.74 Less than 63?  yes.. So you get recycled payout for it.

Something I noticed. Radial Chutes don't amount to as much as Radial
Chutes in Vanilla.  Their area is very small.  In the 120 area for each
one.. So beware.