Update .66 changes
Fixed:
1. Added hidden goals to mini view, while in mini view goals will hide
automatically, which translate to Normal View.  If in normal view and
select Hide Goal it will translate back to mini View when entered.
Applies to optional goals also.

2. Made the custom contract screen much smaller, should help with those
with smaller resolution monitors.

Added:

1. New Custom Mission Contract addition.  you can now select to capture
a asteroid in Custom contracts, and select from a current list of
subjects.  This is a WIP.  At this time the custom contract version
actually pays a little less than the Random Contract Version.  This is
to help balance out Custom Contracts Vs Random Contracts. Still pays out
well though. But the big payouts are still in Regular Contracts.